% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\docType{data}
\name{community}
\alias{community}
\title{community data}
\format{
An object of class \code{data.frame} with 7 rows and 20 columns.
}
\usage{
data(community)
}
\description{
Dataset on Ecuadorian bird communities to be used as examples in the functions
'givemeNMDS' and 'givemeDIVERSITY'. Organized as site (rows) by species
(columns) matrix with count data in each cell.
}
\examples{

comm <- data(community)
givemeNMDS(comm)
givemeDIVERSITY(comm)
}
\keyword{datasets}
