% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givemeDIVERSITY.R
\name{givemeDIVERSITY}
\alias{givemeDIVERSITY}
\title{givemeDIVERSITY}
\usage{
givemeDIVERSITY(community)
}
\description{
This function takes a sites (rows) x species (columns) matrix,
uses the package 'BAT' to perform alpha and beta diversity calculations,
and returns (1) boxplots depicting alpha diversity across sites,
and (2) pairwise estimates for beta diversity across sites and overall
beta diversity for all sites (including relative contributions of
richness and replacement in determining overall beta).
}
