% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givemeNMDS.R
\name{givemeNMDS}
\alias{givemeNMDS}
\title{givemeNMDS}
\usage{
givemeNMDS(community)
}
\description{
This function takes a sites (rows) x species(columns) matrix,
uses the package 'vegan' to perform distance calculations and non-metric
dimensional scaling analysis, and returns an NMDS plot with all sites. Use
vegan functions 'orditorp' and 'ordihull' to display sites and visualize
site relationships, respectively.
}
