% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givemeTREE.R
\name{givemeTREE}
\alias{givemeTREE}
\title{givemeTREE}
\usage{
givemeTREE(traits)
}
\description{
This function takes a species (rows) by traits (columns) matrix,
performs intermediate distance and clustering calculations,
and plots a dendrogram relating species that share similar traits.
}
