% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\docType{data}
\name{traits}
\alias{traits}
\title{functional traits data}
\format{
An object of class \code{data.frame} with 30 rows and 6 columns.
}
\usage{
data(traits)
}
\description{
Dataset on Ecuadorian bird communities to be used as example in the function
'givemeTREE'. Organized as species (rows) by traits (columns) matrix with
binary 0/1 data in each cell (0 = species does not have trait, 1 = species
has trait).
}
\examples{

traits <- data(traits)
givemeTREE(traits)
}
\keyword{datasets}
