% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givemeNMDS.R, R/givemeSTRESS.R
\name{givemeNMDS}
\alias{givemeNMDS}
\title{givemeNMDS}
\usage{
givemeNMDS(community)

givemeNMDS(community)
}
\description{
This function takes a sites (rows) x species(columns) matrix,
uses the package 'vegan' to perform distance calculations and non-metric
dimensional scaling analysis, and returns an NMDS plot with all sites. Use
vegan functions 'orditorp' and 'ordihull' to display sites and visualize
site relationships, respectively.

This function takes a sites (rows) x species(columns) matrix,
uses the package 'vegan' to perform distance calculations and non-metric
dimensional scaling analysis, and returns a stress plot showing how well
your data fit the analysis. Should be used with 'givemeNMDS' to confirm
that your data is appropriate. May return a warning that says stress is
nearly zero -- this is usually occurs with datasets that have lots of
0 values, which is typically expected in large community datasets.
}
